Move looks_like_test_client_call into ast module
